requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-autoslug==1.9.8
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
Markdown==3.3.6
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
python-slugify==5.0.2
pytz==2021.3
sorl-thumbnail==12.7.0
sqlparse==0.4.2
text-unidecode==1.3
zipp==3.7.0
